+++ app/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/tooling/MockedIntegrationFeatures.java
@@ -1,6 +1,5 @@
 package com.ebsolutions.papertrail.financialdataproviderservice.tooling;
 
-import static io.cucumber.junit.platform.engine.Constants.FILTER_TAGS_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME;
@@ -18,7 +17,6 @@ import org.junit.platform.suite.api.Suite;
 @ConfigurationParameters({
     @ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "json:target/cucumber/cucumber.json"),
     @ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "com.ebsolutions.papertrail.financialdataproviderservice"),
-    @ConfigurationParameter(key = FILTER_TAGS_PROPERTY_NAME, value = "@MockIntegration"),
     @ConfigurationParameter(key = PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, value = "true"),
     @ConfigurationParameter(key = JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME, value = "long")
 })
