--- app/src/test/resources/data.sql
@@ -1,4 +1,0 @@
-INSERT INTO app_user (user_id, username, first_name, last_name)
-VALUES
-(1, 'first_user', 'first', 'user'),
-(2, 'second_user', 'second', 'user');
